// CAC/src/app/model/calificacion.model.ts
export class CalificacionModel {

    /**
     *
     */
    constructor() {
    }

    id: number = 0;
    registrosAuditoriaId: number = 0;
    registrosAuditoriaDetalleId: number = 0;
    variableId: number = 0;
    ipsId: number = 0;
    itemId: number = 0;
    nombreItem: string = "";
    calificacion: number = 0;
    observacion: string = "";
    createdBy: string = "";
    createdDate: any = null;
    modifyBy: string = "";
    modifyDate: any = null;

}